refactor: make useNotice hook
알람을 위한 useNotice hook 생성.
추후 좀더 추상화하여 사용할 예정.
