+++ front/src/components/NavigationBar/DefaultNavBtnContainer/Notice/Notice.js
@@ -1,30 +1,20 @@
-import React, { useReducer, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import Alarm from 'components/NavigationBar/DefaultNavBtnContainer/Notice/Alarm';
-import Messages from 'components/MessageList/MessageList';
+import MessageList from 'components/NavigationBar/DefaultNavBtnContainer/Notice/MessageList/MessageList';
 import { faBell } from '@fortawesome/free-solid-svg-icons';
 import { NavImageStyle } from 'components/NavigationBar/DefaultNavBtnContainer/DefaultNavBtnContainer.style';
 import useAlarm from 'hooks/Alarm/useAlarm';
 import { NoticeStyle } from 'components/NavigationBar/DefaultNavBtnContainer/Notice/Notice.style';
 import useCloseClicker from 'hooks/useCloseClicker';
-
-const noticeTypeReducer = (state, action) => {
-  if (action.type === 'messages') {
-    if (action.open) return { isOpen: true, type: 'messages' };
-    return { isOpen: false, type: 'messages' };
-  }
-
-  if (!action.open) return { isOpen: false, type: action.type };
-
-  if (state.isOpen) return state;
-  if (!state.type) return { type: action.type };
-  return { isOpen: true, type: action.type };
-};
+import useNotice from 'hooks/Alarm/useNotice';
 
 export default function Notice() {
   const [alarmList, alarmListDispatch] = useAlarm();
+  const [notice, noticeDispatch] = useNotice();
 
   useEffect(() => {
     if (!alarmList.length) return () => {};
+
     noticeDispatch({ open: true, type: 'alarm' });
     const timeoutToCloseAlarm = setTimeout(() => {
       noticeDispatch({ open: false, type: 'alarm' });
@@ -33,17 +23,16 @@ export default function Notice() {
     return () => {
       clearTimeout(timeoutToCloseAlarm);
     };
-  }, [alarmList, alarmListDispatch]);
-
-  const [notice, noticeDispatch] = useReducer(noticeTypeReducer, {
-    isOpen: false,
-    type: null,
-  });
+  }, [alarmList, alarmListDispatch, noticeDispatch]);
 
   let noticeElement = null;
   if (notice.isOpen) {
     noticeElement =
-      notice.type === 'alarm' ? <Alarm alarmList={alarmList} /> : <Messages />;
+      notice.type === 'alarm' ? (
+        <Alarm alarmList={alarmList} />
+      ) : (
+        <MessageList />
+      );
   }
 
   const openMessages = () => {
